+++ packages/core/src/index.ts
@@ -1,2 +1,11 @@
 export { slidy } from './slidy';
-export type { Options, SlidyInstance, EasingFunc, AnimationFunc, AnimationArgs, PluginFunc, PluginArgs, EventMap } from './types';
+export type {
+    Options,
+    SlidyInstance,
+    EasingFunc,
+    AnimationFunc,
+    AnimationArgs,
+    PluginFunc,
+    PluginArgs,
+    EventMap,
+} from './types';