+++ packages/core/src/lib/utils.ts
@@ -2,7 +2,7 @@ import type { Options, PluginFunc } from '../types';
 
 const { assign, entries } = Object;
 
-const { abs, exp, floor, min, max, round, sign } = Math
+const { abs, exp, floor, min, max, round, sign } = Math;
 
 function clamp(mn: number, val: number, mx: number): number {
     return min(mx, max(mn, val));
@@ -17,13 +17,13 @@ function throttle(
 ): (args: any) => void {
     return th
         ? (args) => {
-            if (!wait) {
-                fn(args);
-                wait = true;
-                clearTimeout(tm);
-                tm = setTimeout(() => (wait = false), ms);
-            }
-        }
+              if (!wait) {
+                  fn(args);
+                  wait = true;
+                  clearTimeout(tm);
+                  tm = setTimeout(() => (wait = false), ms);
+              }
+          }
         : (args) => fn(args);
 }
 
@@ -31,7 +31,7 @@ function loop(
     array: string | any[] | HTMLCollection | Array<Partial<Options>> | PluginFunc[],
     cb: (item: typeof array[number], i: number, array: any) => void
 ) {
-    for (let i = 0; i < array.length; i++) cb(array[i], i, array)
+    for (let i = 0; i < array.length; i++) cb(array[i], i, array);
     return array;
 }