+++ packages/plugins/src/lib/log.ts
@@ -1,14 +1,14 @@
-import type { PluginArgs } from '../types'
+import type { PluginArgs } from '../types';
 
 export function log(params?: any) {
-    console.log('log', params)
+    console.log('log', params);
 
     return ({ node, options, instance }: PluginArgs) => {
-        console.log('log', { node, options, instance })
+        console.log('log', { node, options, instance });
         node.addEventListener('destroy', function destroy() {
             // destroy event
             node.removeEventListener('destroy', destroy);
         });
-        return null
-    }
+        return null;
+    };
 }