+++ packages/plugins/src/types.ts
@@ -1,8 +1,8 @@
-export type { PluginArgs, PluginFunc, EventMap } from '@slidy/core'
+export type { PluginArgs, PluginFunc, EventMap } from '@slidy/core';
 
 export type TimerInstace = {
     play: () => void;
     pause: () => void;
     resume: () => void;
     stop: () => void;
-}
+};