+++ packages/plugins/src/utils/utils.ts
@@ -1,9 +1,6 @@
-function loop<T>(
-    array: ArrayLike<T>,
-    cb: (item: typeof array[number], i: number) => void
-) {
+function loop<T>(array: ArrayLike<T>, cb: (item: typeof array[number], i: number) => void) {
     for (let i = 0; i < array.length; i++) cb(array[i], i);
     return array;
 }
 
-export { loop }
+export { loop };