+++ packages/solid/src/components/Slidy/i18n.ts
@@ -1,1 +1,1 @@
-export { i18nDefaults } from "@slidy/assets/i18n";
+export { i18nDefaults } from '@slidy/assets/i18n';