+++ packages/core/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## 3.6.1 - fix issues & updates
+## 3.7.0 - add `options.plugins`, fix issues & updates
 
+-   [add plugins to iife](https://github.com/Valexr/Slidy/commit/f8b0442273392f5ba4e48cfcb83431e1ebc4ab6a)
+-   [fix default options](https://github.com/Valexr/Slidy/commit/9b2fb4c3e7e48b76feee42340b23caf6717261b8)
+-   [options.plugins: []](https://github.com/Valexr/Slidy/commit/59be47539bd27622079dbd571ee9daa50b995073)
 -   [fix unsnapped loop mode](https://github.com/Valexr/Slidy/commit/49a730394f0d8040e9bb19d792b23c0553555a1d)
 -   [guard loop in mutate](https://github.com/Valexr/Slidy/commit/50b2a6ecb79d5724db3c6b1b357bad86c997dfc6)
 -   [fix snap === 'deck'](https://github.com/Valexr/Slidy/commit/02a0d4868db5f99461147309f09043c34bb3ff58)